refactor(#3037): relocate AppState to crate::app_state (backflow 9→3)
Six service-layer handler/helper modules imported `crate::server::routes::AppState`
(auto_queue/route, auto_queue/phase_gate_violations, onboarding/channel,
onboarding/mod, session_forwarding, dispatched_sessions) — the bulk of the
remaining service→server backflow.

AppState's dependencies are all neutral or below services (config, engine,
eventbus, services::*, sqlx) — it never needs `crate::server`. Move the struct +
impl into a new crate-root module `crate::app_state` (below both server and
services). `crate::server::routes` re-exports it (`pub use`) so every existing
`crate::server::routes::AppState` call site and `AppState { .. }` construction
resolves unchanged. The six service modules now reference `crate::app_state::AppState`.

broadcast_tx/batch_buffer field types switched from `crate::server::ws::*` to
`crate::eventbus::*` (same underlying types via #3251 re-export) so app_state
holds no server path.

Pure relocation — no logic/field/behavior change. Backflow baseline ratcheted
9→3 (remaining: dispatched_sessions auto_queue::activate ×2, internal_api ×1).
