+++ scripts/audit_maintainability/baselines/service_server_backflow.json
@@ -2,29 +2,14 @@
   "schema_version": 1,
   "rule": "service_server_backflow",
   "description": "No-regression baseline for src/services references to crate::server or super::server server-layer modules.",
-  "captured_at": "2026-06-08",
-  "total_count": 14,
+  "captured_at": "2026-06-09",
+  "total_count": 3,
   "files": {
-    "src/services/auto_queue/phase_gate_violations.rs": {
-      "count": 1
-    },
-    "src/services/auto_queue/route.rs": {
-      "count": 1
-    },
     "src/services/discord/internal_api.rs": {
       "count": 1
     },
     "src/services/dispatched_sessions.rs": {
-      "count": 8
-    },
-    "src/services/onboarding/channel.rs": {
-      "count": 1
-    },
-    "src/services/onboarding/mod.rs": {
-      "count": 1
-    },
-    "src/services/session_forwarding.rs": {
-      "count": 1
+      "count": 2
     }
   }
 }
